requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distlib==0.3.7
distro==1.8.0
executing==2.0.1
filelock==3.13.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.2
openai==1.3.8
packaging==23.2
parso==0.8.3
platformdirs==4.0.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
setuptools==67.7.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
termcolor==2.4.0
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
